Backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.115.12
h11==0.16.0
idna==3.10
motor==3.7.1
paho-mqtt==2.1.0
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.6
pydantic_core==2.33.2
pymongo==4.13.1
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
qrcode==8.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uv==0.7.13
uvicorn==0.34.3
